requirements.txt
boto3==1.18.10
botocore==1.21.10
certifi==2021.5.30
charset-normalizer==2.0.3
docker==5.0.0
idna==3.2
jmespath==0.10.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pywin32==227
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.6
websocket-client==1.1.0
